Fix for enemy collisions
Apparently rigidbodies makes detection playout weird because of physics calculations, all I did was change detection to continous instead of discrete and it worked. Added a couple of debugs so I knew when I was reaching some parts of code.

Also tried a bunch of tags and layers. TO-DO: CLEAN LAYERS AND TAGS (Layer masks in Player script included, lots made)